backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi[all]>=0.135.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
